peer: discover nhs protocol address from error indocation too
If an old opennhrp, which does not listen broadcast address, receives
an broadcast registration request it sends an error indication that
the hop count was exceeded. We can use this indication to detect
the NHS protocol address (this assumes we are connected to some other
NHS running new opennhrp so the error indication can be routed back).