server: clear negatives when receiving resolution request
if we just tried resolving the other, and it was down we
have negative request in cache. when it's up again and tries
to resolve us, we won't be able to reply if we have negative
entry in cache. as getting resolution request is proof of
livelyness we just nuke all negatives when such request is
received.