packet: use only NHS and static mappings for NHRP message routing
For NHRP messages, we should not use cached or other entries.

Resolution requests/reply should be always sent via NHS. Several
scenarios require this:

1) If bindings change in NHS, the information is not updated to
   spokes as they keep refreshing the information via direct
   connection (unless the original spoke is rebooted).

2) If spoke-to-spoke shortcut exists, and the other spoke is
   rebooted, it will quite likely do resolution request to the
   other spoke which would try to incorrectly reply using dead
   IPsec SA (since, the SA will be flushed only when the resolution
   reply has been received back, and new ISAKMP-SA has started
   and INITIAL-CONTACT notification has been sent).

Same applies to Purge messages and Error Indication. Registration
(request and reply) should be always sent directly to NHS or back,
which is already honoured by the code.

Traffic Indication is fixed to be sent only to NHCs we are serving
and only if the message is received directly from the NHC. This
ensures that only the first NHS sends Traffic Indication and only
to NHCs it's directly serving to minimize the indication traffic.
This also means that Traffic Indication is not sent if NHS opennhrp
is restarted, and until NHC re-registers (so IPsec should be always
restarted with opennhrp - until proper purging of dynamics is done
on opennhrp exit).