peer: resolution requests that timeout should not be negative cached
Timeout happens only if:
 - none of NHS is reachable
 - major packet loss

In either of cases, we could just retrigger the discovery of protocol
address, if we still want to talk to them. If the protocol address
is really not available, we get NHRP error back from NHS, and in that
case we should negative cache the entry.