fix race condition to stop processing dns requests
ev_timer_again stops the timer if .repeat is zero, but ares_timeout
returns zero timeout if there's events that are already expired.
This becomes issue if we were busy processing something else and
then c-ares timeouts expire meanwhile, we'd end up not handling them.
As a workaround schedule it with small timeout, while this can
postpone timeout handling quite a bit, in practice this should
work well enough.